refactor: move the api url default to core-http-kit and extract readConfig
@authup/kit is general, context-independent utility — the address the
authup server listens on is service context, so API_URL_DEFAULT moves to
@authup/core-http-kit, next to the client whose baseURL it fills. The
authup launcher drops its fallback entirely and sets NUXT_PUBLIC_API_URL
only when the config file names one, leaving the application's own
default in charge.

readConfig moves out of the config module into a dedicated helpers file.

Also removes the database unique-entries component: a one-shot repair for
a historic mysql unique-check failure that full-scanned roles and
permissions on every boot.
