+++ apps/client-web/nuxt.config.ts
@@ -6,7 +6,7 @@
  */
 
 import type { ModuleOptions } from '@authup/client-web-nuxt';
-import { API_URL_DEFAULT } from '@authup/kit';
+import { API_URL_DEFAULT } from '@authup/core-http-kit';
 import path from 'node:path';
 import { defineNuxtConfig } from 'nuxt/config';
 import tailwindcss from '@tailwindcss/vite';
