+++ apps/server-core/src/app/modules/config/index.ts
@@ -6,6 +6,7 @@
  */
 
 export * from './constants.ts';
+export * from './helpers.ts';
 export * from './module.ts';
 export * from './origins.ts';
 export * from './read/index.ts';
