--- apps/server-core/src/components/database/unique-entries/factory.ts
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2025.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-import type { DataSource } from 'typeorm';
-import { DatabaseUniqueEntriesComponent } from './module.ts';
-
-export function createDatabaseUniqueEntriesComponent(dataSource: DataSource) {
-    return new DatabaseUniqueEntriesComponent(dataSource);
-}
