+++ apps/server-core/src/components/index.ts
@@ -5,7 +5,6 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './database/index.ts';
 export * from './event-cleaner/index.ts';
 export * from './oauth2-cleaner/index.ts';
 export * from './types.ts';
