+++ packages/client-web-nuxt/package.json
@@ -32,6 +32,7 @@
     "dependencies": {
         "@authup/access": "^1.0.0-beta.56",
         "@authup/client-web-kit": "^1.0.0-beta.56",
+        "@authup/core-http-kit": "^1.0.0-beta.56",
         "@authup/kit": "^1.0.0-beta.56",
         "@nuxt/kit": "^4.0.0",
         "pathtrace": "^2.2.0",
