+++ packages/core-http-kit/src/client/constants.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2026.
+ * Author Peter Placzek (tada5hi)
+ * For the full copyright and license information,
+ * view the LICENSE file that was distributed with this source code.
+ */
+
+/**
+ * Base URL a client falls back to when the consumer configured none —
+ * the address server-core listens on with its own default port.
+ */
+export const API_URL_DEFAULT = 'http://localhost:3001';
