+++ packages/core-http-kit/src/client/index.ts
@@ -5,6 +5,7 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
+export * from './constants';
 export * from './helper';
 export * from './module';
 export * from './registry';
